Add parser for PSPW
Add a parser for the NWPW PSPW type calculations.

Also add a caveat for the case of PBC=False in StructureData
AiiDA StructureData objects have an attribute 'pbc' to indicate the three boundary conditions for the cell.
If any of these is false, and the user requests a cell be added, an error is thrown as the geometry stanza
is the wrong place to add this in NWChem.