feat: add nwchem.base calculation
This adds a nwchem.base calculation that simply takes a nwchem input
file, thus making it possible to run any calculation through
aiida-nwchem that nwchem can run.
The previous calculation class remains in place and inherits from the
new base class.

We also add a remote folder that can be used to restart from previous
nwchem calculations (to be implemented).
Since nwchem does not bundle the relevant files for restarts in a
subfolder, we will need to create symlinks for individual files based on
their extension, e.g.
  - aiida.db
  - aiida.molvecs
  - aiida.t1amp.0001 , ...
  - aiida.t2amp.0001 , ...