DevOps: Fix the pre-commit
The `pylint` job is moved from a system installation to having the
installation taken care of by `pre-commit` itself. This requires the
dependencies to be specified however.

The `mypy` step is removed since no typing is in place whatsoever.

Any tool configuration in the `pre-commit` config is moved to the
`pyproject.toml`, e.g., for `flynt` and `isort`.