DevOps: Move the source directory into `src/`
This package directory structure is fast becoming the standard in the
Python packaging community. Even though the PyPA is not yet officially
recommending it as the standard, there are various articles out there
outlining the advantages of this structure over having the source folder
at the root of the repository. See also the discussion of the Python
Packaging Authority:

    https://packaging.python.org/en/latest/discussions/src-layout-vs-flat-layout/