DevOps: Install latest supported RabbitMQ version in CI/CD
The CI and CD workflows were installing the `rabbitmq:latest` image
which on the newer Ubuntu releases would install a recent version of
RabbitMQ that is not compatible with `aiida-core` with the default
server configuration. For the tests this is not really a problem since
problems only occur for long running processes over the daemon, but it
does print a warning. We fix to the image `rabbitmq:3.8.14` which is
the last compatible version.